+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.3.1 (Next)
 - Fix GPG keys output path message.
+- Fix encrypted data validation.
 
 ### 1.3.0
 - Bump minimum Ruby version to 2.6.